major refactor, albeit still broken
