+++ tests/test_xarray.py
@@ -1,6 +1,6 @@
 import numpy as np
 import pytest
-from fibsem_tools.io.xr import flip, stt_array, stt_coord
+from fibsem_tools.coordinate import flip, stt_array, stt_coord
 
 
 @pytest.mark.parametrize("length", (1, 10, 100))
